pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["rl_deploy*"]

[project]
name = "spot-rl-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bosdyn-choreography-client>=5.1.0",
    "bosdyn-client>=5.1.0",
    "bosdyn-mission>=5.1.0",
    "bosdyn-orbit>=5.1.0",
    "onnxruntime>=1.23.2",
    "pygame>=2.6.1",
    "pyps4controller>=1.2.5",
    "spatialmath-python>=1.1.15",
]
